make the `REPL` specific code into an extension
